Move modal to inside of layer info card, simplify prop
